Fix build problems on rhel 5.2 and 64-bit encoding bug.
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@699413 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/cluster/Cluster.cpp
@@ -115,7 +115,7 @@ void Cluster::mcastControl(const framing::AMQBody& body, Connection* cptr) {
     mcastEvent(e);
 }
 
-void Cluster::mcastBuffer(const char* data, size_t size, const ConnectionId& connection, size_t id) {
+void Cluster::mcastBuffer(const char* data, size_t size, const ConnectionId& connection, uint32_t id) {
     Event e(DATA, connection, size, id);
     memcpy(e.getData(), data, size);
     QPID_LOG(trace, "MCAST " << e);
